Zotero + commentary improvements
