+++ Angular/src/app/commentary/translation/translation.component.html
@@ -10,4 +10,3 @@
     </div>
   </div>
 </mat-expansion-panel>
-<hr />
